+++ .husky/pre-commit
@@ -1,4 +1,8 @@
-#!/usr/bin/env sh
+#!/bin/bash
+
+# @license Copyright (c) 2003-2022, CKSource Holding sp. z o.o. All rights reserved.
+# For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
+
 . "$(dirname -- "$0")/_/husky.sh"
 
-npm run husky-pre-commit
+yarn lint-staged
