+++ package.json
@@ -145,8 +145,7 @@
     "url": "https://github.com/ckeditor/ckeditor5.git"
   },
   "scripts": {
-    "husky-pre-commit": "lint-staged",
-    "prepare": "husky install && node ./scripts/run-prepare-in-external.js",
+    "postinstall": "husky install && node ./scripts/external-postinstall.js",
     "lint": "eslint --quiet \"**/*.{js,ts}\" --format ./scripts/eslint-formatter.js",
     "stylelint": "stylelint --quiet --allow-empty-input \"packages/**/*.css\" \"docs/**/*.css\"",
     "test": "node --max_old_space_size=4096 node_modules/@ckeditor/ckeditor5-dev-tests/bin/test.js",
